refactor: OmokDatabaseHelper 역할 분리
